requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
google-genai>=0.3.0
pydantic>=2.0.0
pdf2image>=1.17.0
Pillow>=10.0.0